wip: twitter history injection
