+++ src/database/index.ts
@@ -1,4 +1,3 @@
 export * from "./types";
 export * from "./tweets";
-export * from "./twitter-history";
 export * from "./db";
