+++ src/database/tweets.ts
@@ -1,6 +1,6 @@
 import { Tweet } from "../socialmedia/types";
-import { saveTwitterHistory, getTwitterHistory } from "./twitter-history";
 import { logger } from "../logger";
+import { db } from "../database";
 
 export const insertTweet = (username: string, tweet: Tweet): void => {
   try {
@@ -44,18 +44,33 @@ export const insertTweet = (username: string, tweet: Tweet): void => {
     }
 
     // Save the bot's reply
-    saveTwitterHistory({
-      twitter_user_id: username,
-      tweet_id: tweet.new_tweet_id,
-      tweet_text: tweet.new_tweet_text,
-      created_at: new Date().toISOString(),
-      is_bot_tweet: 1,
-      conversation_id: tweet.conversation_id,
-      prompt: tweet.prompt,
-      username: username,
-    });
+    const stmt = db.prepare(`
+      INSERT INTO twitter_history (
+        twitter_user_id,
+        tweet_id,
+        tweet_text,
+        created_at,
+        is_bot_tweet,
+        conversation_id,
+        prompt,
+        username,
+        input_tweet_id
+      ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)
+    `);
 
-    logger.info("Successfully inserted tweet");
+    stmt.run(
+      username,
+      tweet.new_tweet_id,
+      tweet.new_tweet_text,
+      new Date().toISOString(),
+      1,
+      tweet.conversation_id,
+      tweet.prompt,
+      username,
+      tweet.input_tweet_id,
+    );
+
+    logger.debug("Successfully inserted tweet");
   } catch (e) {
     logger.error("Error inserting tweet:", e);
     if (e instanceof Error) {
@@ -67,19 +82,31 @@ export const insertTweet = (username: string, tweet: Tweet): void => {
 
 export const getTweetByInputTweetId = (id: string): Tweet | undefined => {
   try {
-    const history = getTwitterHistory(id, 1);
-    if (!history.length) return undefined;
+    logger.debug(`Checking for tweet ID: ${id}`);
+    
+    const stmt = db.prepare(`
+      SELECT * FROM twitter_history 
+      WHERE input_tweet_id = ?
+      LIMIT 1
+    `);
+    const tweet = stmt.get(id) as TwitterHistory;
+    
+    logger.debug('Query result:', JSON.stringify(tweet, null, 2));
+    
+    if (!tweet) {
+      logger.debug('No tweet found');
+      return undefined;
+    }
 
-    const tweet = history[0];
     return {
-      input_tweet_id: tweet.tweet_id,
+      input_tweet_id: tweet.input_tweet_id || tweet.tweet_id,
       input_tweet_created_at: tweet.created_at,
       input_tweet_text: tweet.tweet_text,
       input_tweet_user_id: tweet.twitter_user_id,
       input_tweet_username: tweet.username,
-      new_tweet_id: "",
+      new_tweet_id: tweet.tweet_id,
       prompt: tweet.prompt || "",
-      new_tweet_text: "",
+      new_tweet_text: tweet.tweet_text,
       conversation_id: tweet.conversation_id,
     };
   } catch (e) {
@@ -113,5 +140,212 @@ export const getLastTweetByUsername = (username: string): Tweet | undefined => {
       logger.error("Error stack:", e.stack);
     }
     return undefined;
+  }
+};
+
+export interface TwitterHistory {
+  twitter_user_id: string;
+  tweet_id: string;
+  tweet_text: string;
+  created_at: string;
+  is_bot_tweet: number;
+  conversation_id?: string;
+  prompt?: string;
+  username?: string;
+  input_tweet_id?: string;
+}
+
+export const saveTwitterHistory = (history: TwitterHistory) => {
+  try {
+    logger.debug("Saving twitter history");
+
+    // Validate required fields
+    if (
+      !history.twitter_user_id ||
+      !history.tweet_id ||
+      !history.tweet_text ||
+      !history.created_at
+    ) {
+      throw new Error(
+        `Missing required fields: ${JSON.stringify({
+          twitter_user_id: !history.twitter_user_id,
+          tweet_id: !history.tweet_id,
+          tweet_text: !history.tweet_text,
+          created_at: !history.created_at,
+        })}`,
+      );
+    }
+
+    const stmt = db.prepare(`
+      INSERT INTO twitter_history (
+        twitter_user_id,
+        tweet_id,
+        tweet_text,
+        created_at,
+        is_bot_tweet,
+        conversation_id,
+        prompt,
+        username,
+        input_tweet_id
+      ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)`);
+
+    stmt.run(
+      history.twitter_user_id,
+      history.tweet_id,
+      history.tweet_text,
+      history.created_at,
+      history.is_bot_tweet,
+      history.conversation_id || null,
+      history.prompt || null,
+      history.username || null,
+      history.input_tweet_id || null,
+    );
+
+    logger.debug("Successfully saved twitter history");
+  } catch (e) {
+    logger.error(
+      "Error saving twitter history:",
+      e instanceof Error ? e.message : e,
+    );
+    if (e instanceof Error) {
+      logger.error("Error stack:", e.stack);
+    }
+    throw e;
+  }
+};
+
+export const getTwitterHistory = (
+  userId: string,
+  limit: number = 50,
+  conversationId?: string,
+): TwitterHistory[] => {
+  try {
+    let query = `
+      SELECT * FROM twitter_history 
+      WHERE twitter_user_id = ?
+    `;
+    const params: any[] = [userId];
+
+    if (conversationId) {
+      query += ` AND conversation_id = ?`;
+      params.push(conversationId);
+    }
+
+    query += ` ORDER BY created_at DESC LIMIT ?`;
+    params.push(limit);
+
+    return db.prepare(query).all(...params) as TwitterHistory[];
+  } catch (e) {
+    logger.error(`Error getting twitter history for user ${userId}:`, e);
+    if (e instanceof Error) {
+      logger.error("Error stack:", e.stack);
+    }
+    return [];
+  }
+};
+
+export const getTwitterHistoryByUsername = (
+  username: string,
+  limit: number = 50,
+): TwitterHistory[] => {
+  try {
+    return db
+      .prepare(
+        `SELECT * FROM twitter_history WHERE username = ? ORDER BY created_at DESC LIMIT ?`,
+      )
+      .all(username, limit) as TwitterHistory[];
+  } catch (e) {
+    logger.error(`Error getting twitter history for user ${username}:`, e);
+    if (e instanceof Error) {
+      logger.error("Error stack:", e.stack);
+    }
+    return [];
+  }
+};
+
+export const getConversationHistory = (
+  conversationId: string,
+  limit: number = 50,
+): TwitterHistory[] => {
+  try {
+    return db
+      .prepare(
+        `SELECT * FROM twitter_history WHERE conversation_id = ? ORDER BY created_at DESC LIMIT ?`,
+      )
+      .all(conversationId, limit) as TwitterHistory[];
+  } catch (e) {
+    logger.error(
+      `Error getting conversation history for ${conversationId}:`,
+      e,
+    );
+    if (e instanceof Error) {
+      logger.error("Error stack:", e.stack);
+    }
+    return [];
+  }
+};
+
+export const formatTwitterHistoryForPrompt = (
+  history: TwitterHistory[],
+  includePrompts: boolean = false,
+): string => {
+  try {
+    return history
+      .map(tweet => {
+        let text = `@${tweet.username}: ${tweet.tweet_text}`;
+        if (includePrompts && tweet.prompt) {
+          text += `\nPrompt used: ${tweet.prompt}`;
+        }
+        return text;
+      })
+      .join("\n\n");
+  } catch (e) {
+    logger.error("Error formatting twitter history:", e);
+    if (e instanceof Error) {
+      logger.error("Error stack:", e.stack);
+    }
+    return "";
+  }
+};
+
+export const getUserInteractionCount = (
+  twitterUserId: string,
+  botUsername: string,
+  interactionTimeout: number,
+): number => {
+  try {
+    const cutoff = new Date(Date.now() - interactionTimeout).toISOString();
+
+    const result = db
+      .prepare(
+        `SELECT COUNT(*) as interaction_count 
+         FROM twitter_history 
+         WHERE twitter_user_id = ? 
+         AND is_bot_tweet = 1
+         AND created_at >= ?
+         AND (
+           -- Count when we reply to this user
+           (username = ?) OR
+           -- Count when this user is in the conversation
+           (conversation_id IN (
+             SELECT conversation_id 
+             FROM twitter_history 
+             WHERE twitter_user_id = ?
+             AND conversation_id IS NOT NULL
+           ))
+         )`,
+      )
+      .get(botUsername, cutoff, twitterUserId, twitterUserId) as { interaction_count: number };
+
+    return result.interaction_count;
+  } catch (e) {
+    logger.error(
+      `Error getting interaction count for user ${twitterUserId}:`,
+      e,
+    );
+    if (e instanceof Error) {
+      logger.error("Error stack:", e.stack);
+    }
+    return 0;
   }
 };
